Fix incorrect snapshotting of MRP config by CASE client. (#32380)
* Fix incorrect snapshotting of MRP config by CASE client.

Right now, we snapshot the MRP config as part of the CASEClientInitParams during stack startup.

After that, we will use that snapshotted config whenever we are the CASE initiator.

This config will not match the parameters we use as CASE responder or advertise
over DNS-SD if the local MRP configuration ever changes.  Which for an ICD it can.

The fix is to stop the incorrect snapshotting and get the information we need
from the right source of truth when we need it.

* Address review comments.

* Address issue with CASE session not assuming the right things when NullOptional is passed in.